Added the beginning of the test domain and test runner. Started implementation of the AutoRunner. Ported and adapted assembly resolver manager.
Added support for AssemblyResolverAttribute and IAssemblyResolver to the
MbUnitTestFramework.  This is done in a flexible way by instrumenting
test assembly loading instead of having specific support in the core for
the feature.  In general I've tried to keep the core fully decoupled from
the framework.

Ported the MbUnit.AddIn project over.  It's currently commented out so
the build doesn't fail.

Ported the command-line parsing utilities from MbUnit v2.
Added command-line option handling to MbUnit Echo and linked to the AutoRunner.

Reworked the reporting service to decouple Xml report production from the
reporting API itself by using an IReportWriter as intermediary.  This
greatly simplifies the implementation and opens up other possibilities.

Added Castle Windsor as an IoC.  It's not fully integrated yet but it's the
direction I intend to take to link together all of the runtime components.
I wasn't 100% behind this decision at first because in my opinion a test
runner probably shouldn't have too many dependencies.  However, an IoC
container is really what's needed here and I'm certainly not going to go
implement my own!

A hundred misc other changes...  Still very much a work in progress and
it is changing fluidly.

The MbUnit runner doesn't get very far yet but it loads test assemblies into
an isolated AppDomain.  That's ok, I'll refine the process with tests.  The
important thing for me was to get the general shape of things in place so
I'd have a large enough vocabulary to think with when writing the tests.