Turned the framework inside-out!  *phew*
Now MbUnit.Framework contains all API for test writers and MbUnit.Core
depends on MbUnit.Framework.

Reverted the introduction of ILMerge, this arrangement provides a much better
way to solve the problem of proliferating dependencies in test assemblies.

NOTE: This is only a partial check-in.  SVN can't reliably move files that have
just been added or modified.  So I still need to do that to rearrange the projects.