Things are starting to get interesting...
Refactored the ITestFramework to be represented as an ITestHarnessContributor
over an ITestHarness.  ITestHarness captures the test context including the
list of loaded assemblies, a reference to the runtime, and all of the
behavior associated with building the template and test trees and running
tests.  The old TestProject class is gone and instead we pass around references
to ITestHarness.

Renamed TestTemplate* classes to Template*.  I got tired of that extra word
in there which wasn't really accurate anyways since a Template might not always
be specifying a test.  Updated all language to talk about template trees instead
of test template trees.

Refactored the based model object underneath template and test trees.  Now
we've got IModelComponent, ITestComponent and ITemplateComponent.  This makes
things clearer, particularly as regards the MetadataPatternAttribute which
appeared to be somewhat ambiguous as to which tree it was operating on
(templates, of course but still).

Added support to MbUnit.Echo to save the template and test trees to XML.
This is useful for diagnostic purposes among other things.

Extended the AutoRunner to provide references to the template and test trees
on demand.

Lots of other misc. changes.

As soon as I finalize the test tree and the transformations that take place
in MbUnit to populate the test tree, we'll be able to begin work on Icarus
in earnest and open up the core.  That's true even if we can't self-host
yet because at least the object model will be complete.