It runs MbUnit v2 tests! Really! That much closer to self-hosting and having stable external APIs...
Plumbed in more of the remaining bits of the test execution pipeline.
The idea is that we group tests into batches that are managed by individual
controllers.  The controllers can then run them however they like and
report back what they've done via an IEventListener.  We marshal these
events across test domains so that the test runner can see them.

Cleaned up ITestRunner somewhat.
Introduced a little separation of concerns in the form of test runner monitors
to capture things like writing to the console.

Renamed "Report" to "ExecutionLog" because it's clearer that way.

Moved all of the test harness code back into the Core assembly.  The framework
shouldn't ever need to know this stuff exists.

Updated the MbUnit v2 assemblies to an interim version that allows Gallio
to run MbUnit v2 tests (just needed to make one method virtual).  Also did a
bunch of work on the MbUnit v2 plugin including grabbing metadata and actually
running the tests and collecting the results.  Seems to work ok.

Portions of this checkin are still subject to change.  I expect another big
refactoring is due...

Changed the metadata map to only allow string values.  I'll be adding a
different kind of decoration for declarative information that is invisible to
the test runner.