Applied patch from Dave Griffiths
Added validity for outlying cases. For messages and formatted messages (with args), I've put in asserts to ensure
that the message does actually get displayed. 

Also noticed a bit of dodgy code in Asserts, looks like some Resharper madness, but if called it would have caused a recursive infinite-loop, and
also causes an ambiguous invocation compiler error. I've supplied a patch to remove it. Also added an argument check inside 2 of AreEqual() methods (the
ones with delta arguments), as a negative value would have caused the Assert to act incorrectly.