Commit work in progress building the reporting API and some refactoring.
Added IsTestCase to make it possible to distinguish tests from fixtures
and other things programmatically.  Relying on the Kind metadata would be a
bad idea for this property.

Fleshed out FilterParser.

Removed the bogus dependency on MbUnit.Echo from MbUnit.Gallio.Core.Tests.

Note: Not done yet.  The tests pass but that's only because I haven't
      integrated the new work yet.  ;-)