Integrated the Xml report formatter.
Applied some cleanups and fixes to the test runners based on TestRunnerHelper to eliminate some unnecessary code and enable report generation.

Improved the handling of cancelation during an Echo test run.  It will generate a partial report.

Added IProgressMonitor.ThrowIfCanceled which polls IsCanceled and throws OperationCanceledException when appropriate.  This can simplify handling of cancelation upstream.

Added support for loading and saving reports with the option of embedding attachments in the Xml file or saving them all separately.