Refactored and improved the console support somewhat more.  The main idea was to make the use of a shared console less error-prone and to ease testing concerns somewhat.  So I've introduced an IRichConsole interface that encapsulates the console features we care about and is amenable to dependency injection.
Also improved the progress monitoring story a little bit by adding an IProgressMonitorProvider
to capture the whole lifecycle of a progress monitor including its creation and disposal.
The provider can also choose to run the task in another thread if needed.  That'll be useful
for progress monitor dialogs.

Some other minor tweaks to TestRunnerHelper.  Removed an unnecessary convenience constructor
and made the filter setting optional.