Having trouble focusing on stuff today so I attacked a couple minor nits.
Moved out of the way some tentative API for assertions I wrote last year.  I'm not ready to commit my work in progress quite yet.

Breaking API change: Moved AssertionException to the MbUnit.Framework namespace.  Note this change improves compatibility with NUnit whose AssertionException is in NUnit.Framework.  Of course users probably shouldn't be writing tests that manipulate AssertionExceptions directly.

Fixed the NAnt tests to go find the NAnt.exe that lives in the project's libs folder instead of looking at Program Files.

Optimized the RichConsoleProgressMonitor to reduce flickering somewhat.

Modified DebugMonitor to write debug messages to an ILogger.  Now we can see real-time debug output when tests run with Debug Verbosity.  (Probably not something you want to do often but it's useful for diagnostics...)

Refactored the command-line parsing code a bit so that the printing of the response file argument and the default argument goes through the same path as the rest of the argument printing.  This ensures consistent output and word-wrapping.  Previously the response file argument description wasn't indented correctly.
Also fixed the printing of the default argument description.
Cleaned up and externalized the error messages and other text.
Added support for a silent "/?" synonym for help in Echo.
Replaced magic numbers in the command line output class with constants.
Removed the command line utility as it really wasn't all that useful.
Made the command line argument parser stateless so it can be reused without doing all of the reflection again.
Clarified the control flow and data flow of the command line argument parser somewhat.  I found it a little hard to follow because of the use of instance fields isntead of more tightly scoped local variables and parameters.

Cleaned up Echo's main Run method and got rid of SetUp.  I think it's easier to read now that the intermediate "haltExecution" and "resultCode" fields are gone.
Command line argument parsing errors go to the console instead of the logger.  It's a bit clearer.
Eliminated redundant printing of the full path of assemblies that couldn't be located, it's not that useful.