Major refactoring and significant new work.
Introduced the ITestPlan abstraction for scheduling test execution order.  An ITestPlan works together with ITestControllers to run tests.  Test execution state is tracked by ITestMonitor and IStepMonitor objects that are provided by the ITestPlan.  Basically the ITestPlan feeds a sorted tree of ITestMonitor objects to the appropriate controllers.  The controllers then take care of running the test steps associated with each test and providing result codes.

Added support for Contexts.  Eliminated the clunky interfaces and left out some convenience methods planned for later.

Added support for Steps via the Context object.

Reworked the execution log API to simplify it down to a few abstract classes.  The interfaces are gone.

Improved the separation of concerns between the Core and Framework assemblies.  Moved a lot of stuff out of the Framework that only the Core really needs to know.  Also combined and rearranged the namespaces.  (I should run NDepend over Gallio to see how it looks now.  I'm sure the module structure can still be improved.)

Changed the order in which declarative information is processed when generating templates.  Also clearly documented what's going on in each of the Apply() methods.  One major change was to ensure that members declared by supertypes are processed before members declared by subtypes.  That property is crucial for ensuring that situations such as classes that have multiple [TestFixtureSetUp] methods are processed correctly.

Renamed TestState to TestStatus (affects XSLT for reports).

Added support for reading the XML documentation associated with any code element.  This is all fully tested, including gnarly cases involving generic method parameters and nested types.  All 3 supported test frameworks now expose XML documentation as the XmlDocumentation metadata element of tests.

Changed the MbUnit.Echo default to echo results to the console.  Use the /ne switch to turn that off.  I figure it makes more sense to show too much information to a first-time user than none at all.  The problem was that if you ran some tests and didn't specify /e or /rt you would see the final test summary but you'd have no way to see the actual failures.  Bad...

Added some new lifecycle phases.

Eliminated the IExecutionLogWriter type.  Instead execution logs are based on the abstract LogWriter and LogStreamWriter classes.  Rearranged some of the API to make it more concise (for example, no more Streams property).  Added an EmbedExisting method to embed an existing attachment given only its id.  The LogWriter and LogStreamWriter classes are completely composable so that's how RemoteLogWriter is implemented now.  The LogEventType and LogEventArgs types were also updated for obvious reasons.  Attaching an attachment for the first time is now a distinct operation from embedding it.  Furthermore in the case of a RemoteLogWriter, we now guarantee that we never send an attachment more than once over the wire for a given test step.  We also perform stricter validation of the attachments to ensure we don't allow multiple attachments with the same id.

There are tons of changes and bug fixes but the main new feature is this:

We can now run some Gallio tests for real.  It's almost dogfood!