After seeing the test tree in Icarus I decided to simplify and clean up the nodes that are generated by each of the test frameworks supported so they are more uniform.  Keep in mind that namespace and other information is available by looking at the CodeReference or Metadata of the tests.
Fixed an obscure bug in MetadataMap's Xml deserialization (and added a regression test).

Found that in a number of places I was doing some magic with Assembly.CodeBase and Assembly.Location so I consolidated that into RuntimeUtils.

Added the CodeBase assembly-level metadata to the report.

Cleaned up the dependency between Context and Runtime.

Fixed a few broken tests.