I was inspired by NDepend to do a bit more refactoring and housekeeping in the Gallio codebase.  It happened that I had a few unwanted dependencies in places that should not be mutually dependent.
Fixed problems running the MSBuild and NAnt task unit tests within Gallio
that were caused by the fact we were trying to reinitialize the runtime
within the scope of a running test.

Significantly reworked the runtime initialization process and introduced
stronger guards to prevent accidental reinitialization and attempts
to use an uninitialized runtime.

The runtime is global.  Just deal with it.  Previously I'd been passing
IRuntime instances around all over the place.  Not only was this a nuisance
but it promised something that could not be achieved: supporting multiple
simultaneous runtime instances.  The dependency injection didn't accomplish
anything that couldn't already be done by initializing the runtime appropriately.

Added a runtime setup option to specify a runtime factory very early on.
This can be useful if a host application wants to use something other
than Windsor as a container.

Cleaned up the AssemblyResolverManager story a little bit and removed the
explicit dependence on MbUnit.

Extracted a new high-level Loader class for accessing assembly load related services.

Moved the Mono flag to the Runtime.

Integrated the VS debug helper as part of the DefaultAssemblyResolverManager.

Got rid of ITestHarnessContributor.  It was too general and vague.

Cut down ITestFramework to bare essentials.

Added an ITestEnvironment to capture environmental configuration concerns
in an extensible manner.  Refactored TestEnvironment to become the
ConsoleTestEnvironment and TraceTestEnvironment components that only handle
Console and Trace concerns respectively.

Merged MbUnit.Model.Events into MbUnit.Model.Execution since it's only relevant for execution
anyways and there were mutual dependencies here.

Moved TestPackage up to MbUnit.Model to eliminate mutual dependencies.

Refactored the ITestRunner classes and renamed them more consistently.
simplified the initialization strategy for LocalTestRunner and StandaloneTestRunner
and introduced a new IsolatedTestRunner (which is probably what Icarus will
be using soon).

Renamed TestRunnerHelper to TestLauncher.  Moved result-interpretation related
concerns to TestLauncherResult.  Did a bunch of work to bring the TestLauncher
up to the standard of a full-blown API and made it a little bit more flexible
(primarily to simplify test concerns related to other components that use it.)
Its purpose and usage should be much more obvious now.

Renamed a bunch of test fixtures to end with the non-pluralized word "Test"
for consistency.

More dogfooding of the unit tests.

Various other minor changes.