Fixed problems running NAnt and MSBuild integration tests from the command-line build.  They were failing because the Gallio plugins could not be located by the task after a Release build.
Refactored the NAnt and MSBuild integration tests to highlight their commonalities.

Extracted a generally useful ProcessRunner class and added it to the MbUnit framework.  It captures console output and error streams and prints them to the log in a distinct section along with the final exit code.  I imagine this API will evolve somewhat over time.

Fixed a bug in the NAnt task's handling of plugin and hint directories.  It was looking at DirSet.FileNames rather than DirSet.DirectoryNames.

Removed the Readme.txt in the TestBuildFiles folders.  I think the contents are sufficiently self-explanatory.

Updated my previous hack for disabling compiler inlining of certain critical methods (that must appear in a stack trace) to use a NonInlinedAttribute.  That makes the hack easier to find at least but I still don't like it.  There's got to be a better way...

Added a couple more tests.