Bit of refactoring and more progress on the ReSharper test runner.
Added code to initialize the Gallio runtime environment in the context of the remote task runner.  As with the RuntimeProxy itself, the trick is that we need to avoid referencing any Gallio types along this critical path until initialization takes place.  In particular, this means that the remote task objects cannot contain reference to Gallio types because they are instantiated before the runner has a chance to set things up.

Tests still don't actually run yet.