** Big redesign of the test object model. ** (Work in progress)
Dropped templates in favor of parameterized tests that have zero or more test instances created at runtime.
Added support for generic parameters.
Added data binding primitives.
Redesigned the PatternAttribute mechanism to be more composable.  Now there are only 3 methods that must be defined on an IPattern interface.  PatternAttribute subclasses supply template method implementations of common attribute types.
Switch to using the ITestExplorer API exclusively for test enumeration.
Refactored TestModel and TestPackage.
Added an ErrorTest.
Added the ReSharper test runner plugin to the installer.
Added the PowerShell cmdlet to the build scripts and to the installer.
Added a few framework reference libraries for .Net 3.5 featurs and the PowerShell cmdlet.
Refactored Context and IContextManager.
Refactored the ITestMonitor and ITestPlan implementations to simplify all of the synchronization logic and the relationship between monitors and the test Context.
Added an MbUnit.Compatibility project for providing backwards compatible implementations of features that have changed from MbUnit v2 to v3 to ease migration.
Added skeleton implementations of a bunch of new MbUnit attributes.
Added more features to the reflection abstraction layer as needed.

Known bugs:
- Residual problems with the framework adapters to be fixed shortly.
- The test reports don't know how to handle multiple testInstanceRun elements correctly (but some progress has been made to update the XSL to take into account reporting XML schema changes)
- The installer does not register the PowerShell cmdlet yet.
- Incomplete testing overall.