Today I learned that the IsNullable property of Xml attributes actually controls whether a null element is omitted or whether it is represented by an xsi:null="true" attribute.  Accordingly, I decided that those extra elements were quite useful to have in the Xml report schema so I've set IsNullable to false everywhere.  This still looks strange to me since the flag doesn't actually talk about whether a null value is permitted, only how a null is represented.  Oh well, I'd best get used to it.
In keeping with the theme of stripping irrelevant junk out of reports, made the CodeReference property of TestComponentData nullable.  This way we avoid having a bunch of junk code reference elements in there for nothing.  It does mean that clients now have to deal with null values in addition to CodeReference.Unknown values but that's easily dealt with using the ?? null coalescing operator.  Let's see how it goes, I might change my mind again later.