Added some tolerance to the debug symbol tests to handle the fact that Debug and Release builds may contain different sequence points due to optimization so the line numbers might vary a little.
Fixed a bug in the NAnt test due to it being unable to create a mock of an internal type.  It turns out we didn't really need the extra interface anyways since the InstrumentedGallioTask can just as easily stub out the Log methods itself by overridding them.

Renamed the logger classes to follow the same convention used for the PowerShell cmdlet.  Namely, the logger is named based on what it logs to, in this case it logs to an MSBuild or NAnt task so we'll call them TaskLoggers.