Replaced the ValueFormatter with a RuleBasedFormatter. Added quite a few rules to handle common formatting cases well, including things like enumerations, dictionary entries and XPathNavigables. Added tests for StringUtils. Added some new helpers to StringUtils. Added missing license boilerplate.
Note: Found a bug related to data binding with generic parameters.  If the type of one of the slots is itself a generic parameter, we are unable to resolve what its value should be when it comes time to perform value type conversions.  I'm going to sleep on it and fix it tomorrow.  Quite likely I'll add a generic parameter translation feature to DataBindingContext that can accept constant type substitutions as well as other IDataBindingAccessors defined by the context.  It may require a few safety checks to avoid recursive cycles but it shouldn't be too bad.  It certainly seems better to keep the complexity of the resolution and conversion process inside DataBindingContext.